adaly-test/backend-test/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.18
alembic==1.11.1
pydantic==2.0.2
pydantic-settings==2.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.6
pytest==7.3.1
httpx==0.24.1